pyproject.toml
[project]
name = "litterbot-wrapped"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pylitterbot>=2025.0.0",
    "python-dotenv>=1.2.1",
]
